% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidity.R
\name{checkValidity}
\alias{checkValidity}
\title{A checkValidity Function}
\usage{
checkValidity(dat, numeric = FALSE, cat = FALSE)
}
\arguments{
\item{dat}{Data in data frame format.}

\item{numeric}{Numeric data type.}

\item{cat}{Categorical or factor data type.}
}
\value{
Returns a list of numeric and character(factor) variables, basic summary statistics and number of rows and columns in the data frame
}
\description{
'checkValidity' checks the class of the different variables in the data frame and returns a list
of numeric and factor or character variables. It also returns basic summary statistics(mean, median, min, max) for numeric and categorical/character variables
It also returns the number of colunms and rows in the data frame.
}
\examples{
library(FactoMineR)
data(wine)
checkValidity(wine, numeric=TRUE)
}
\author{
Henry Nanji, Saisakul Chernbumroong
}
\keyword{datatype}
