% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctType.R
\name{correctType}
\alias{correctType}
\title{A correctType Function}
\usage{
correctType(raw, datatype, cols = NULL)
}
\arguments{
\item{raw}{Data in data frame format.}

\item{datatype}{The data type for each raw data column. Accepted data types are factor, numeric, integer,
date, ordered, and character.}

\item{cols}{If not all columns need correction, specify the column number that need correction.}
}
\value{
Returns the data set with the corrected data type.
}
\description{
'correctType' reads data from .csv file into the correct data type.
This function processes the raw data into the correct data type.
}
\examples{
library(FactoMineR)
data(wine)
correctType(wine, "factor", 1)
}
\author{
Henry Nanji, Saisakul Chernbumroong
}
\keyword{data}
\keyword{type}
