% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualInfo.R
\name{mutualInfo}
\alias{mutualInfo}
\title{A mutualInfo Function}
\usage{
mutualInfo(x, y)
}
\arguments{
\item{x}{Numerical or factor data.}

\item{y}{Numerical or factor data.}
}
\value{
Returns the mutual information value.
}
\description{
'mutualInfo' function calculates mutual information of the two variables.
}
\examples{
data(iris)
mutualInfo(iris[, 1], iris[, 5])
}
\author{
Henry Nanji, Saisakul Chernbumroong
}
\keyword{information}
\keyword{mutual}
