% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outlier_detection}
\alias{outlier_detection}
\title{Outlier detection function}
\usage{
outlier_detection(dat, ...)
}
\arguments{
\item{dat}{A data frame for data values.}

\item{...}{Other arguments.}
}
\value{
Returns a box plot showing the outliers for each variable.
}
\description{
'outlier_detection'visually detect and highlights outliers in a univariate continuous variable.
The function fetches the values of data points that lie beyond the extremes of the whiskers(observations that lie outside of 1.5 * IQR.
}
\examples{
library(FactoMineR)
data(wine)
outlier_detection(wine)
}
\author{
Henry  Nanji
}
\keyword{outlier}
