% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotGroup.R
\name{plotBoxplotGroup}
\alias{plotBoxplotGroup}
\title{A plotBoxplotGroup Function}
\usage{
plotBoxplotGroup(dat, x, ...)
}
\arguments{
\item{dat}{Data in frame format.}

\item{x}{A grouping variable.}

\item{...}{Other parameters.}
}
\value{
Returns a box plot for each grouping variable
}
\description{
'plotBoxplotGroup' function plots boxplots by group.
}
\examples{
data(iris)
plotBoxplotGroup(iris, 'Species')
}
\author{
Henry Nanji
}
\keyword{box}
\keyword{plot}
