% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapGroup.R
\name{plotHeatmapGroup}
\alias{plotHeatmapGroup}
\title{A plotHeatmapGroup function for numeric variables}
\usage{
plotHeatmapGroup(dat, ...)
}
\arguments{
\item{dat}{Data in data frame format.}

\item{...}{Optional arguments}
}
\value{
A heatmap is returned showing correlation of all numeric variables in data frame
}
\description{
'plotHeatmapGroup' returns a plot showing the strength of correlation of numeric variables in a data frame.
 The function receives just one input as data frame.
}
\examples{
data(iris)
plotHeatmapGroup(iris[,1:4])
}
\author{
Henry  Nanji,Saisakul Chernbumroong
}
\keyword{heatmap}
